+++ src/components/App/App.tsx
@@ -1,8 +1,8 @@
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
 import {AppRoute} from '../../const';
-import MainPage from '../../pages/MainPage/MainPage';
-import LoginPage from '../../pages/LoginPage/LoginPage';
-import FavoritesPage from '../../pages/FavoritesPage/FavoritesPage';
+import MainPage from '../../pages/main-page/main-page';
+import LoginPage from '../../pages/login-page/login-page';
+import FavoritesPage from '../../pages/favorites-page/favorites-page';
 import OfferPage from '../../pages/offer-page/offer-page';
 import NotFoundPage from '../../pages/not-found-page/not-found-page';
 import PrivateRoute from '../private-route';
